Module Syguslib

module Expressions : sig ... end

Convenience module for writing expressions within predefined theories.

module Parser : sig ... end

This module defines function to convert from s-expressions to sygus.

module Semantic : sig ... end

This module defines a few high-level properties of SyGuS programs. Setter commands and some utility functions on solver responses are specified here.

module Serializer : sig ... end

A Sygus command or a sygus term is an s-expression. This modules defines functions to convert from a SyGuS term to an s-expression. The functions in this module should be self-explanatory.

module Solvers : sig ... end

module Sygus : sig ... end

This module contains definitions that mirror the SyGuS Language Standard Version 2.1, which can be found at https://org/assets/pdf/SyGuS-IF_2.1.pdf.