Module Syguslib.Expressions

Convenience module for writing expressions within predefined theories.

val var : Sygus.symbol -> Sygus.sygus_term

Create a term containing a variable given a variable name.

val sort : Sygus.symbol -> Sygus.sygus_sort

Create a sort given a sort name.

val int : int -> Sygus.sygus_term

Create an integer constant term given an int.

val real : float -> Sygus.sygus_term

Create a decimal number term given a float.

val hex : string -> Sygus.sygus_term

Create a hex constant given its string representation.

val ite : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Create a conditional given the condition, then-branch and else-branch expressions.

val (/) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Numeral division.

val (+) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Numeral addition.

val (-) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Numeral Substraction.

val (*) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Numeral multiplication.

val max : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Max function (non-standard, needs to be declared first in a SyGuS file.)

val min : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Min function (non-standard, needs to be declared first in a SyGuS file.)

val modulo : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Modulo operation.

val abs : Sygus.sygus_term -> Sygus.sygus_term

Abs operations.

val neg : Sygus.sygus_term -> Sygus.sygus_term

Unary minus.

val mk_false : Sygus.sygus_term

The false value.

val mk_true : Sygus.sygus_term

The true value in Sygus.

val (&&) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Conjunction.

val (||) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

Disjunction.

val not : Sygus.sygus_term -> Sygus.sygus_term

Negation.

val (>) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

val (>=) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

val (<) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

val (<=) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

val (=) : Sygus.sygus_term -> Sygus.sygus_term -> Sygus.sygus_term

val gconst : Sygus.sygus_sort -> Sygus.sygus_gsterm

gconst x is the sygus grammar term (Constant x)

val gterm : Sygus.sygus_term -> Sygus.sygus_gsterm

gterm x is the sygus grammar term x

val gvar : Sygus.sygus_sort -> Sygus.sygus_gsterm

gvar x is the sygus grammar term (Variable x)

val gblock : string -> Sygus.sygus_sort -> Sygus.sygus_gsterm list -> Sygus.grammar_def

glblock n s productions defines the grammar productions with name n and sort s with the productions productions.

val int_sort : Sygus.sygus_sort

The standard sort for integers.

val bool_sort : Sygus.sygus_sort

The standard sort for booleans.

val real_sort : Sygus.sygus_sort

The standard sort for reals.