Module Syguslib.Semantic

This module defines a few high-level properties of SyGuS programs. Setter commands and some utility functions on solver responses are specified here.

Semantic identity of commands

val is_setter_command : Sygus.command -> Base.bool

Returns true when the command is a setter command, that is, sets some option. logic, or feature for the solver.

val is_well_formed : Sygus.program -> Base.bool

Returns true if the program is well-formed. A well-formed program has the form (set-logic ..)(setter-command ..)*(other-commands..)*

val declares : Sygus.command -> Base.string Base.list

Returns the list of symbols declared by a command, in no particular order

val compare_declares : Sygus.command -> Sygus.command -> Base.int

Compare two commands by the symbols they declare.

Static definitions

val max_definition : Sygus.command

A static definition for the max operation.

val min_definition : Sygus.command

A static definition for the min operation.

Transformers

val rename : (Base.string * Base.string) Base.list -> Sygus.sygus_term -> Sygus.sygus_term

A transformation function that renames all variables in a sygus term.

I/O functions

val write_command : Stdio.Out_channel.t -> Sygus.command -> Base.unit

Writes a command to an output channel.

Wrapper modules for important types

module Command : sig ... end

module Term : sig ... end

module Ident : sig ... end

module Lit : sig ... end

module Sort : sig ... end