Module Solvers.SyncSolver

A module encapsulating function to execute SyGuS solvers synchronously.

Parameters

module Stats : Statistics

module Log : Logger

module Config : SolverSystemConfig

Signature

module CoreSolver : sig ... end

val kill_solver : int -> unit

Kill a solver, given its process id. Ignores errors.

val exec_solver : ?solver_kind:CoreSolver.t -> ?options:string list -> ?pid:int Stdlib.ref -> ?error_log:string option -> (string * string) -> Sygus.solver_response

Execute a solver on a given sygus set of commands. exec_solver ~solver_kind ~options ~pid ~error_log (inputfile, outputfile) executes the solver ~solver_kind (defaults to CVC) with options ~options and stores the process id of the solver process in ~pid as soon as the process is started. The solver is executed by printing the commands in inputfile and outputting the solver response in outputfile.

val solve_commands : ?solver_kind:CoreSolver.t -> ?pid:int Stdlib.ref -> Sygus.program -> Sygus.solver_response

solve_commands ~solver_kind ~pid program solvers the set of commands in program using the solver solver_kind (default to CVC) and stores the process id of the solver in pid. The user can use pid as a handle to kill the process if necessary.