Module type Solvers.SolverSystemConfig

A module to provide the system configuration for syntax-guided synthesis solver. It indicates the path to the solver executables on the system. CVC4 and CVC5 are treted as one solver, with the boolean use_cvc5 () setting which version of CVC to use.

val cvc_binary_path : unit -> string

val dryadsynth_binary_path : unit -> string

val eusolver_binary_path : unit -> string

val using_cvc5 : unit -> bool